refactor: split index file into separate modules, one per step
